+++ tests/browser/async_returnvalue.js
@@ -5,6 +5,12 @@ mergeInto(LibraryManager.library, {
         wakeUp(value + 1);
       }, 1);
     });
+  },
+  sync_tunnel_bool: function(value) {
+    return Asyncify.handleSleep(function(wakeUp) {
+      setTimeout(function() {
+        wakeUp(!value);
+      }, 1);
+    });
   }
 });
-