+++ tests/pthread/test_pthread_c11_threads.c
@@ -3,8 +3,10 @@
 #include <assert.h>
 #include <threads.h>
 
+tss_t key;
 once_flag flag = ONCE_FLAG_INIT;
 _Atomic int counter = 0;
+_Atomic int destructor_counter = 0;
 
 void do_once(void) {
   counter++;
@@ -13,6 +15,7 @@ void do_once(void) {
 
 int thread_main(void* arg) {
   printf("in thread_main %p\n", thrd_current());
+  tss_set(key, thrd_current());
   call_once(&flag, do_once);
   printf("done thread_main\n");
   return 42;
@@ -23,9 +26,17 @@ int run_with_exit(void* arg) {
   thrd_exit(43);
 }
 
+void destructor(void* val) {
+  printf("destructor: %p\n", thrd_current());
+  assert(val == thrd_current());
+  destructor_counter++;
+}
+
 int main(int argc, char* argv[]) {
   int result = 0;
   printf("thrd_current: %p\n", thrd_current());
+
+  assert(tss_create(&key, destructor) == thrd_success);
 
   // Test call_once
   thrd_t t1;
@@ -48,6 +59,8 @@ int main(int argc, char* argv[]) {
   assert(result == 42);
   assert(counter == 1);
 
+  assert(destructor_counter == 4);
+
   // Test thrd_exit return value
   thrd_t t5;
   assert(thrd_create(&t5, run_with_exit, NULL) == thrd_success);
@@ -58,6 +71,7 @@ int main(int argc, char* argv[]) {
   thrd_t t6;
   assert(thrd_create(&t6, thread_main, NULL) == thrd_success);
   assert(thrd_detach(t6) == thrd_success);
+
 
 #ifdef REPORT_RESULT
   REPORT_RESULT(0);