+++ tests/test_core.py
@@ -99,6 +99,9 @@ def with_both_exception_handling(f):
         self.skipTest('wasm2js does not support wasm exceptions')
       if not config.V8_ENGINE or config.V8_ENGINE not in config.JS_ENGINES:
         self.skipTest('d8 required to run wasm eh tests')
+      # FIXME Temporarily disabled. Enable this later when the bug is fixed.
+      if '-fsanitize=address' in self.emcc_args:
+        self.skipTest('Wasm EH does not work with asan yet')
       self.emcc_args.append('-fwasm-exceptions')
       self.v8_args.append('--experimental-wasm-eh')
       self.js_engines = [config.V8_ENGINE]
@@ -164,6 +167,7 @@ def also_with_standalone_wasm(wasm2c=False, impure=False):
           # support in order for JS to have a chance to run this without trapping
           # when it sees an i64 on the ffi.
           self.set_setting('WASM_BIGINT')
+          self.emcc_args.append('-Wno-unused-command-line-argument')
           # if we are impure, disallow all wasm engines
           if impure:
             self.wasm_engines = []
@@ -748,18 +752,18 @@ class TestCoreBase(RunnerCore):
       }
     ''')
 
-    building.emcc('a1.c', ['-c'])
-    building.emcc('a2.c', ['-c'])
-    building.emcc('b1.c', ['-c'])
-    building.emcc('b2.c', ['-c'])
-    building.emcc('main.c', ['-c'])
+    self.emcc('a1.c', ['-c'])
+    self.emcc('a2.c', ['-c'])
+    self.emcc('b1.c', ['-c'])
+    self.emcc('b2.c', ['-c'])
+    self.emcc('main.c', ['-c'])
 
     building.emar('cr', 'liba.a', ['a1.c.o', 'a2.c.o'])
     building.emar('cr', 'libb.a', ['b1.c.o', 'b2.c.o'])
 
     building.link_to_object(['main.c.o', 'liba.a', 'libb.a'], 'all.o')
 
-    building.emcc('all.o', self.get_emcc_args(), 'all.js')
+    self.emcc('all.o', self.get_emcc_args(), 'all.js')
     self.do_run('all.js', 'result: 1', no_build=True)
 
   def test_if(self):
@@ -1819,11 +1823,10 @@ int main() {
   def test_emscripten_get_compiler_setting(self):
     src = test_file('core', 'emscripten_get_compiler_setting.c')
     output = shared.unsuffixed(src) + '.out'
-    old = self.get_setting('ASSERTIONS', 1)
     # with assertions, a nice message is shown
     self.set_setting('ASSERTIONS')
     self.do_runf(src, 'You must build with -s RETAIN_COMPILER_SETTINGS=1', assert_returncode=NON_ZERO)
-    self.set_setting('ASSERTIONS', old)
+    self.clear_setting('ASSERTIONS')
     self.set_setting('RETAIN_COMPILER_SETTINGS')
     self.do_runf(src, open(output).read().replace('waka', shared.EMSCRIPTEN_VERSION))
 
@@ -2411,6 +2414,14 @@ The current type of b is: 9
   def test_pthread_dispatch_after_exit(self):
     self.do_run_in_out_file_test('pthread', 'test_pthread_dispatch_after_exit.c')
 
+  @node_pthreads
+  def test_pthread_thread_local_storage(self):
+    self.set_setting('PROXY_TO_PTHREAD')
+    self.set_setting('EXIT_RUNTIME')
+    self.set_setting('PTHREAD_POOL_SIZE', 8)
+    self.set_setting('INITIAL_MEMORY', '300mb')
+    self.do_run_in_out_file_test('pthread', 'test_pthread_thread_local_storage.cpp')
+
   def test_tcgetattr(self):
     self.do_runf(test_file('termios', 'test_tcgetattr.c'), 'success')
 
@@ -2614,8 +2625,6 @@ The current type of b is: 9
   @needs_dylink
   def test_dlfcn_missing(self):
     self.set_setting('MAIN_MODULE')
-    if self.has_changed_setting('ASSERTIONS'):
-      self.skipTest('test needs to customize ASSERTIONS')
     self.set_setting('ASSERTIONS')
     src = r'''
       #include <dlfcn.h>
@@ -3583,7 +3592,7 @@ ok
     return self.dylink_testf(main, side, expected, force_c, **kwargs)
 
   def dylink_testf(self, main, side, expected=None, force_c=False, main_emcc_args=[],
-                   need_reverse=True, auto_load=True, main_module=2, **kwargs):
+                   need_reverse=True, auto_load=True, **kwargs):
     self.maybe_closure()
     # Same as dylink_test but takes source code as filenames on disc.
     old_args = self.emcc_args.copy()
@@ -3594,6 +3603,7 @@ ok
 
     # side settings
     self.clear_setting('MAIN_MODULE')
+    self.clear_setting('ERROR_ON_UNDEFINED_SYMBOLS')
     self.set_setting('SIDE_MODULE')
     side_suffix = 'wasm' if self.is_wasm() else 'js'
     if isinstance(side, list):
@@ -3605,13 +3615,18 @@ ok
     shutil.move(out_file, 'liblib.so')
 
     # main settings
-    self.set_setting('MAIN_MODULE', main_module)
+    self.set_setting('MAIN_MODULE', 2)
     self.clear_setting('SIDE_MODULE')
     if auto_load:
       self.emcc_args += main_emcc_args
       self.emcc_args.append('liblib.so')
-      if force_c:
-        self.emcc_args.append('-nostdlib++')
+    else:
+      # When not auto loading (i.e. when not specifying the side module on the
+      # command line) we need to disable warnings on undefined symbols.
+      self.set_setting('WARN_ON_UNDEFINED_SYMBOLS', 0)
+
+    if force_c:
+      self.emcc_args.append('-nostdlib++')
 
     if isinstance(main, list):
       # main is just a library
@@ -3628,7 +3643,7 @@ ok
       # Test the reverse as well.  There we flip the role of the side module and main module.
       # - We add --no-entry since the side module doesn't have a `main`
       self.dylink_testf(side, main, expected, force_c, main_emcc_args + ['--no-entry'],
-                        need_reverse=False, main_module=main_module, **kwargs)
+                        need_reverse=False, **kwargs)
 
   def do_basic_dylink_test(self, **kwargs):
     self.dylink_test(r'''
@@ -3886,7 +3901,6 @@ ok
 
   @needs_dylink
   def test_dylink_i64(self):
-    # Runs with main_module=1 due to undefined getTempRet0 otherwise
     self.dylink_test(r'''
       #include <stdio.h>
       #include <stdint.h>
@@ -3900,12 +3914,11 @@ ok
       int64_t sidey() {
         return 42;
       }
-    ''', 'other says 42.', force_c=True, main_module=1)
+    ''', 'other says 42.', force_c=True)
 
   @all_engines
   @needs_dylink
   def test_dylink_i64_b(self):
-    # Runs with main_module=1 due to undefined getTempRet0 otherwise
     self.dylink_test(r'''
       #include <stdio.h>
       #include <stdint.h>
@@ -3935,12 +3948,11 @@ ok
         x = 18 - x;
         return x;
       }
-    ''', 'other says -1311768467750121224.\nmy fp says: 43.\nmy second fp says: 43.', force_c=True, main_module=1)
+    ''', 'other says -1311768467750121224.\nmy fp says: 43.\nmy second fp says: 43.', force_c=True)
 
   @needs_dylink
   @also_with_wasm_bigint
   def test_dylink_i64_c(self):
-    # Runs with main_module=1 due to undefined getTempRet0 otherwise
     self.dylink_test(r'''
       #include <stdio.h>
       #include <inttypes.h>
@@ -3988,7 +4000,7 @@ res64 - external 64\n''', header='''
       #include <stdint.h>
       EMSCRIPTEN_KEEPALIVE int32_t function_ret_32(int32_t i, int32_t j, int32_t k);
       EMSCRIPTEN_KEEPALIVE int64_t function_ret_64(int32_t i, int32_t j, int32_t k);
-    ''', force_c=True, main_module=1)
+    ''', force_c=True)
 
   @needs_dylink
   @also_with_wasm_bigint
@@ -4240,11 +4252,10 @@ res64 - external 64\n''', header='''
 
     test('libc++')
     test('1')
-    if not self.has_changed_setting('ASSERTIONS'):
-      self.set_setting('ASSERTIONS', 0)
-      test('', expect_pass=False, need_reverse=False)
-      self.set_setting('ASSERTIONS')
-      test('', expect_pass=False, need_reverse=False)
+    self.set_setting('ASSERTIONS', 0)
+    test('', expect_pass=False, need_reverse=False)
+    self.set_setting('ASSERTIONS')
+    test('', expect_pass=False, need_reverse=False)
 
   @needs_dylink
   @with_env_modify({'EMCC_FORCE_STDLIBS': 'libc++'})
@@ -4316,8 +4327,6 @@ res64 - external 64\n''', header='''
   @with_both_exception_handling
   @needs_dylink
   def test_dylink_raii_exceptions(self):
-    # MAIN_MODULE=1 still needed in this test due to:
-    # https://github.com/emscripten-core/emscripten/issues/13786
     self.dylink_test(main=r'''
       #include <stdio.h>
       extern int side();
@@ -4344,14 +4353,13 @@ res64 - external 64\n''', header='''
         volatile ifdi p = func_with_special_sig;
         return p(2.18281, 3.14159, 42);
       }
-    ''', expected=['special 2.182810 3.141590 42\ndestroy\nfrom side: 1337.\n'], main_module=1)
+    ''', expected=['special 2.182810 3.141590 42\ndestroy\nfrom side: 1337.\n'])
 
   @needs_dylink
   @disabled('https://github.com/emscripten-core/emscripten/issues/12815')
   def test_dylink_hyper_dupe(self):
     self.set_setting('INITIAL_MEMORY', '64mb')
-    if not self.has_changed_setting('ASSERTIONS'):
-      self.set_setting('ASSERTIONS', 2)
+    self.set_setting('ASSERTIONS', 2)
 
     # test hyper-dynamic linking, and test duplicate warnings
     create_file('third.cpp', r'''
@@ -4419,10 +4427,9 @@ res64 - external 64\n''', header='''
                      # in wasm, we can't flip as the side would have an EM_ASM, which we don't support yet TODO
                      need_reverse=not self.is_wasm())
 
-    if not self.has_changed_setting('ASSERTIONS'):
-      print('check warnings')
-      full = self.run_js('src.js')
-      self.assertContained("warning: symbol '_sideg' from '%s' already exists" % libname, full)
+    print('check warnings')
+    full = self.run_js('src.js')
+    self.assertContained("warning: symbol '_sideg' from '%s' already exists" % libname, full)
 
   @needs_dylink
   def test_dylink_load_compiled_side_module(self):
@@ -4460,8 +4467,9 @@ res64 - external 64\n''', header='''
 
   @needs_dylink
   def test_dylink_dso_needed(self):
-    def do_run(src, expected_output):
-      self.do_run(src + 'int main() { return test_main(); }', expected_output)
+    def do_run(src, expected_output, emcc_args=[]):
+      create_file('main.c', src + 'int main() { return test_main(); }')
+      self.do_runf('main.c', expected_output, emcc_args=emcc_args)
     self._test_dylink_dso_needed(do_run)
 
   @needs_dylink
@@ -4586,12 +4594,9 @@ res64 - external 64\n''', header='''
     }
     '''
 
-    # MAIN_MODULE=1 still needed in this test due to:
-    # https://github.com/emscripten-core/emscripten/issues/13786
     self.dylink_test(main=main,
                      side=side,
                      header=header,
-                     main_module=1,
                      expected='success')
 
   @needs_dylink
@@ -5091,7 +5096,6 @@ main( int argv, char ** argc ) {
   def test_stat_mknod(self):
     self.do_runf(test_file('stat', 'test_mknod.c'), 'success')
 
-  @no_safe_heap('https://github.com/emscripten-core/emscripten/issues/12433')
   def test_fcntl(self):
     self.add_pre_run("FS.createDataFile('/', 'test', 'abcdef', true, true, false);")
     self.do_run_in_out_file_test('fcntl', 'test_fcntl.c')
@@ -5864,6 +5868,8 @@ return malloc(size);
     # Improves test readability
     self.emcc_args.append('-Wno-c++11-narrowing')
     self.emcc_args.extend(['-Wpedantic', '-Werror', '-Wall', '-xc++'])
+    # Ignore deprecation errors for now
+    self.emcc_args.append('-Wno-error=deprecated-declarations')
     run()
     self.emcc_args.append('-funsigned-char')
     run()
@@ -5886,8 +5892,7 @@ return malloc(size);
     self.run_process([shared.CLANG_CXX, src, '-msse', '-o', 'test_sse1', '-D_CRT_SECURE_NO_WARNINGS=1'] + clang_native.get_clang_native_args(), stdout=PIPE)
     native_result = self.run_process('./test_sse1', stdout=PIPE).stdout
 
-    orig_args = self.emcc_args
-    self.emcc_args = orig_args + ['-I' + test_file('sse'), '-msse']
+    self.emcc_args += ['-I' + test_file('sse'), '-msse']
     self.maybe_closure()
 
     self.do_runf(src, native_result)
@@ -5902,8 +5907,7 @@ return malloc(size);
     self.run_process([shared.CLANG_CXX, src, '-msse2', '-Wno-argument-outside-range', '-o', 'test_sse2', '-D_CRT_SECURE_NO_WARNINGS=1'] + clang_native.get_clang_native_args(), stdout=PIPE)
     native_result = self.run_process('./test_sse2', stdout=PIPE).stdout
 
-    orig_args = self.emcc_args
-    self.emcc_args = orig_args + ['-I' + test_file('sse'), '-msse2', '-Wno-argument-outside-range']
+    self.emcc_args += ['-I' + test_file('sse'), '-msse2', '-Wno-argument-outside-range']
     self.maybe_closure()
     self.do_runf(src, native_result)
 
@@ -5915,8 +5919,7 @@ return malloc(size);
     self.run_process([shared.CLANG_CXX, src, '-msse3', '-Wno-argument-outside-range', '-o', 'test_sse3', '-D_CRT_SECURE_NO_WARNINGS=1'] + clang_native.get_clang_native_args(), stdout=PIPE)
     native_result = self.run_process('./test_sse3', stdout=PIPE).stdout
 
-    orig_args = self.emcc_args
-    self.emcc_args = orig_args + ['-I' + test_file('sse'), '-msse3', '-Wno-argument-outside-range']
+    self.emcc_args += ['-I' + test_file('sse'), '-msse3', '-Wno-argument-outside-range']
     self.maybe_closure()
     self.do_runf(src, native_result)
 
@@ -5928,8 +5931,7 @@ return malloc(size);
     self.run_process([shared.CLANG_CXX, src, '-mssse3', '-Wno-argument-outside-range', '-o', 'test_ssse3', '-D_CRT_SECURE_NO_WARNINGS=1'] + clang_native.get_clang_native_args(), stdout=PIPE)
     native_result = self.run_process('./test_ssse3', stdout=PIPE).stdout
 
-    orig_args = self.emcc_args
-    self.emcc_args = orig_args + ['-I' + test_file('sse'), '-mssse3', '-Wno-argument-outside-range']
+    self.emcc_args += ['-I' + test_file('sse'), '-mssse3', '-Wno-argument-outside-range']
     self.maybe_closure()
     self.do_runf(src, native_result)
 
@@ -5942,8 +5944,7 @@ return malloc(size);
     self.run_process([shared.CLANG_CXX, src, '-msse4.1', '-Wno-argument-outside-range', '-o', 'test_sse4_1', '-D_CRT_SECURE_NO_WARNINGS=1'] + clang_native.get_clang_native_args(), stdout=PIPE)
     native_result = self.run_process('./test_sse4_1', stdout=PIPE).stdout
 
-    orig_args = self.emcc_args
-    self.emcc_args = orig_args + ['-I' + test_file('sse'), '-msse4.1', '-Wno-argument-outside-range']
+    self.emcc_args += ['-I' + test_file('sse'), '-msse4.1', '-Wno-argument-outside-range']
     self.maybe_closure()
     self.do_runf(src, native_result)
 
@@ -5955,8 +5956,7 @@ return malloc(size);
     self.run_process([shared.CLANG_CXX, src, '-msse4.2', '-Wno-argument-outside-range', '-o', 'test_sse4_2', '-D_CRT_SECURE_NO_WARNINGS=1'] + clang_native.get_clang_native_args(), stdout=PIPE)
     native_result = self.run_process('./test_sse4_2', stdout=PIPE).stdout
 
-    orig_args = self.emcc_args
-    self.emcc_args = orig_args + ['-I' + test_file('sse'), '-msse4.2', '-Wno-argument-outside-range']
+    self.emcc_args += ['-I' + test_file('sse'), '-msse4.2', '-Wno-argument-outside-range']
     self.maybe_closure()
     self.do_runf(src, native_result)
 
@@ -5968,8 +5968,7 @@ return malloc(size);
     self.run_process([shared.CLANG_CXX, src, '-mavx', '-Wno-argument-outside-range', '-o', 'test_avx', '-D_CRT_SECURE_NO_WARNINGS=1'] + clang_native.get_clang_native_args(), stdout=PIPE)
     native_result = self.run_process('./test_avx', stdout=PIPE).stdout
 
-    orig_args = self.emcc_args
-    self.emcc_args = orig_args + ['-I' + test_file('sse'), '-mavx', '-Wno-argument-outside-range']
+    self.emcc_args += ['-I' + test_file('sse'), '-mavx', '-Wno-argument-outside-range']
     self.maybe_closure()
     self.do_runf(src, native_result)
 
@@ -6112,10 +6111,11 @@ return malloc(size);
         '-Wno-shift-negative-value',
         '-Wno-format'
     ]
-    asserts = self.get_setting('ASSERTIONS', 0)
 
     # extra testing for ASSERTIONS == 2
-    self.set_setting('ASSERTIONS', 2 if use_cmake else asserts)
+    if use_cmake:
+      self.set_setting('ASSERTIONS', 2)
+      self.emcc_args.append('-Wno-unused-command-line-argument')
 
     self.do_runf(test_file('third_party', 'bullet', 'Demos', 'HelloWorld', 'HelloWorld.cpp'),
                  [open(test_file('bullet', 'output.txt')).read(), # different roundings
@@ -6371,7 +6371,6 @@ return malloc(size);
     'minimal_runtime': ['-s', 'MINIMAL_RUNTIME=1']
   })
   def test_dyncall_specific(self, *args):
-    emcc_args = self.emcc_args.copy()
     cases = [
         ('DIRECT', []),
         ('DYNAMIC_SIG', ['-s', 'DYNCALLS=1', '-s', 'DEFAULT_LIBRARY_FUNCS_TO_INCLUDE=$dynCall']),
@@ -6385,18 +6384,14 @@ return malloc(size);
 
     for which, extra_args in cases:
       print(str(args) + ' ' + which)
-      self.emcc_args = emcc_args + ['-D' + which] + list(args) + extra_args
-      self.do_core_test('dyncall_specific.c')
+      self.do_core_test('dyncall_specific.c', emcc_args=['-D' + which] + list(args) + extra_args)
 
   def test_getValue_setValue(self):
     # these used to be exported, but no longer are by default
     def test(output_prefix='', args=[], assert_returncode=0):
-      old = self.emcc_args.copy()
-      self.emcc_args += args
       src = test_file('core', 'getValue_setValue.cpp')
       expected = test_file('core', 'getValue_setValue' + output_prefix + '.out')
-      self.do_run_from_file(src, expected, assert_returncode=assert_returncode)
-      self.emcc_args = old
+      self.do_run_from_file(src, expected, assert_returncode=assert_returncode, emcc_args=args)
 
     # see that direct usage (not on module) works. we don't export, but the use
     # keeps it alive through JSDCE
@@ -6419,13 +6414,10 @@ return malloc(size);
         if use_files:
           args += ['-DUSE_FILES']
         print(args)
-        old = self.emcc_args.copy()
-        self.emcc_args += args
         self.do_runf(test_file('core', 'FS_exports.cpp'),
                      (open(test_file('core', 'FS_exports' + output_prefix + '.out')).read(),
                       open(test_file('core', 'FS_exports' + output_prefix + '_2.out')).read()),
-                     assert_returncode=assert_returncode)
-        self.emcc_args = old
+                     assert_returncode=assert_returncode, emcc_args=args)
 
       # see that direct usage (not on module) works. we don't export, but the use
       # keeps it alive through JSDCE
@@ -6608,10 +6600,9 @@ return malloc(size);
     # ensure function names are preserved
     self.emcc_args += ['--profiling-funcs']
     self.do_core_test('test_demangle_stacks.cpp', assert_returncode=NON_ZERO)
-    if not self.has_changed_setting('ASSERTIONS'):
-      print('without assertions, the stack is not printed, but a message suggesting assertions is')
-      self.set_setting('ASSERTIONS', 0)
-      self.do_core_test('test_demangle_stacks_noassert.cpp', assert_returncode=NON_ZERO)
+    print('without assertions, the stack is not printed, but a message suggesting assertions is')
+    self.set_setting('ASSERTIONS', 0)
+    self.do_core_test('test_demangle_stacks_noassert.cpp', assert_returncode=NON_ZERO)
 
   def test_demangle_stacks_symbol_map(self):
     # disable aggressive inlining in binaryen
@@ -7040,7 +7031,7 @@ someweirdtext
     no_maps_filename = 'no-maps.out.js'
 
     assert '-gsource-map' not in self.emcc_args
-    building.emcc('src.cpp', self.get_emcc_args(), out_filename)
+    self.emcc('src.cpp', self.get_emcc_args(), out_filename)
     # the file name may find its way into the generated code, so make sure we
     # can do an apples-to-apples comparison by compiling with the same file name
     shutil.move(out_filename, no_maps_filename)
@@ -7049,10 +7040,10 @@ someweirdtext
     no_maps_file = re.sub(' *//[@#].*$', '', no_maps_file, flags=re.MULTILINE)
     self.emcc_args.append('-gsource-map')
 
-    building.emcc(os.path.abspath('src.cpp'),
-                  self.get_emcc_args(),
-                  out_filename,
-                  stderr=PIPE)
+    self.emcc(os.path.abspath('src.cpp'),
+              self.get_emcc_args(),
+              out_filename,
+              stderr=PIPE)
     map_referent = out_filename if not self.is_wasm() else wasm_filename
     # after removing the @line and @sourceMappingURL comments, the build
     # result should be identical to the non-source-mapped debug version.
@@ -7061,24 +7052,7 @@ someweirdtext
     # optimizer can deal with both types.
     map_filename = map_referent + '.map'
 
-    def encode_utf8(data):
-      if isinstance(data, dict):
-        for key in data:
-          data[key] = encode_utf8(data[key])
-        return data
-      elif isinstance(data, list):
-        for i in range(len(data)):
-          data[i] = encode_utf8(data[i])
-        return data
-      elif isinstance(data, type(u'')):
-        return data.encode('utf8')
-      else:
-        return data
-
     data = json.load(open(map_filename))
-    if str is bytes:
-      # Python 2 compatibility
-      data = encode_utf8(data)
     if hasattr(data, 'file'):
       # the file attribute is optional, but if it is present it needs to refer
       # the output file.
@@ -7092,9 +7066,6 @@ someweirdtext
     mappings = json.loads(self.run_js(
       path_from_root('tools', 'source-maps', 'sourcemap2json.js'),
       args=[map_filename]))
-    if str is bytes:
-      # Python 2 compatibility
-      mappings = encode_utf8(mappings)
     seen_lines = set()
     for m in mappings:
       if m['source'] == 'src.cpp':
@@ -7128,7 +7099,7 @@ someweirdtext
     js_filename = 'a.out.js'
     wasm_filename = 'a.out.wasm'
 
-    building.emcc('src.cpp', self.get_emcc_args(), js_filename)
+    self.emcc('src.cpp', self.get_emcc_args(), js_filename)
 
     out = self.run_process([shared.LLVM_DWARFDUMP, wasm_filename, '-all'], stdout=PIPE).stdout
 
@@ -7525,6 +7496,12 @@ Module['onRuntimeInitialized'] = function() {
     self.set_setting('ASYNCIFY_IMPORTS', ['suspend'])
     self.set_setting('ASSERTIONS')
     self.do_core_test('asyncify_assertions.cpp')
+
+  def test_asyncify_during_exit(self):
+    self.set_setting('ASYNCIFY')
+    self.set_setting('ASSERTIONS')
+    self.set_setting('EXIT_RUNTIME', 1)
+    self.do_core_test('asyncify_during_exit.cpp', assert_returncode=1)
 
   @no_asan('asyncify stack operations confuse asan')
   @no_wasm2js('TODO: lazy loading in wasm2js')
@@ -8488,7 +8465,7 @@ NODEFS is no longer included by default; build with -lnodefs.js
     self.set_setting('TOTAL_STACK', 4 * 1024 * 1024)
     self.do_core_test('test_stack_get_free.c')
 
-  # Tests Settings.ABORT_ON_WASM_EXCEPTIONS
+  # Tests settings.ABORT_ON_WASM_EXCEPTIONS
   def test_abort_on_exceptions(self):
     self.set_setting('ABORT_ON_WASM_EXCEPTIONS')
     self.set_setting('EXPORTED_RUNTIME_METHODS', ['ccall', 'cwrap'])
@@ -8499,6 +8476,12 @@ NODEFS is no longer included by default; build with -lnodefs.js
   def test_gl_main_module(self):
     self.set_setting('MAIN_MODULE')
     self.do_runf(test_file('core', 'test_gl_get_proc_address.c'))
+
+  @needs_dylink
+  def test_main_module_js_symbol(self):
+    self.set_setting('MAIN_MODULE', 2)
+    self.emcc_args += ['--js-library', test_file('core', 'test_main_module_js_symbol.js')]
+    self.do_runf(test_file('core', 'test_main_module_js_symbol.c'))
 
   def test_REVERSE_DEPS(self):
     create_file('connect.c', '#include <sys/socket.h>\nint main() { return (int)&connect; }')
@@ -8523,6 +8506,10 @@ def make_run(name, emcc_args, settings=None, env=None):
     env = {}
   if settings is None:
     settings = {}
+  if settings:
+    # Until we create a way to specify link-time settings separately from compile-time settings
+    # we need to pass this flag here to avoid warnings from compile-only commands.
+    emcc_args.append('-Wno-unused-command-line-argument')
 
   TT = type(name, (TestCoreBase,), dict(run_name=name, env=env, __module__=__name__))  # noqa